backend/pyproject.toml
[project]
name = "sortlens-backend"
version = "0.3.5"
description = "Sortlens - Local-first photo management backend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
    "pydantic-settings>=2.1.0",
    "aiosqlite>=0.19.0",
    "pillow>=10.2.0",
    "send2trash>=1.8.2",
    "watchdog>=4.0.0",
    "rawpy>=0.19.0",
    "numpy>=1.26.0",
    "ImageHash>=4.3.1",
    "exifread>=3.0.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.26.0",
    "ruff>=0.2.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
